+++ threedigrid_builder/__init__.py
@@ -2,5 +2,5 @@ from .application import *  # NOQA
 from .exceptions import *  # NOQA
 
 # fmt: off
-__version__ = '0.15.1.dev0'
+__version__ = '0.16.0'
 # fmt: on