+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 0.15.1 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Handle non-ASCII characters in gridadmin HDF5 output.
 
 
 0.15.0 (2022-01-05)